+++ CHANGELOG.md
@@ -38,6 +38,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 - Add CRs to create a bastion host in CAPZ cluster template.
 - Enable termination events for CAPZ node pools.
+- Add support for creating workload cluster keypairs using the `login` command.
 
 ## [1.41.1] - 2021-10-04